This PR includes the following changes:
1. Moves TOML build parsing into a separate module.
2. Corrects the ordering of `apply_*_config` method calls.
3. Moves download logic specific to `toml_config` into its own module.
4. Updates tests to align with the correct behavior.
