+++ src/bootstrap/src/core/config/tests.rs
@@ -9,14 +9,15 @@ use build_helper::git::PathFreshness;
 use clap::CommandFactory;
 use serde::Deserialize;
 
+use super::Config;
 use super::flags::Flags;
 use super::toml::change_id::ChangeIdWrapper;
-use super::{Config, RUSTC_IF_UNCHANGED_ALLOWED_PATHS};
 use crate::ChangeId;
 use crate::core::build_steps::clippy::{LintConfig, get_clippy_rules_in_order};
 use crate::core::build_steps::llvm;
 use crate::core::build_steps::llvm::LLVM_INVALIDATION_PATHS;
 use crate::core::config::toml::TomlConfig;
+use crate::core::config::toml::rust::RUSTC_IF_UNCHANGED_ALLOWED_PATHS;
 use crate::core::config::{LldMode, Target, TargetSelection};
 use crate::utils::tests::git::git_test;
